requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
djangorestframework==3.15.2
docker==7.1.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.3
pytest-django==4.9.0
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
